Fix pytest fixture not getting picked up
pytest fixture was in incorrect location, causing error if "test" database was not present,
lacking clear explanation of how to fix.
Fix by moving to conftest.py https://docs.pytest.org/en/stable/fixture.html#conftest-py-sharing-fixture-functions
